requirements.txt
python-kis>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
anthropic>=0.40.0
pykrx>=1.0.45
